refactor(agents): extract worktree isolation
Move child workspace and git worktree lifecycle helpers behind a narrow module boundary. Keep request parsing at the orchestration edge and avoid coupling the new module to SpawnRequest.
